# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestão de EPIs.",
        "title": "SaaS EPI API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte API",
            "url": "http://www.radaptech.com.br",
            "email": "suporte@seusaas.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna o resumo para o dashboard da empresa",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Empresas"
                ],
                "summary": "Dashboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResumoDashboard"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/empresa": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cadastra uma nova empresa no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Empresas"
                ],
                "summary": "Criar empresa",
                "parameters": [
                    {
                        "description": "Dados da empresa",
                        "name": "empresa",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EmpresaInserir"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Empresa criada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/empresa/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAut": []
                    }
                ],
                "description": "Atualiza os dados de uma empresa existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Empresas"
                ],
                "summary": "Editar empresa",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da empresa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualização",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EditarEmpresaRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sem conteúdo (Sucesso)"
                    },
                    "400": {
                        "description": "ID ou dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/empresas": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna uma lista com todas as empresas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Empresas"
                ],
                "summary": "Listar empresas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Empresa"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/empresas/recentes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista as empresas adicionadas recentemente",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Empresas"
                ],
                "summary": "Empresas recentes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.EmpresaRecente"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "helper.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.EditarEmpresaRequest": {
            "type": "object",
            "required": [
                "email",
                "nome",
                "planoId",
                "responsavel",
                "status"
            ],
            "properties": {
                "cnpj": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "observacoes": {
                    "type": "string"
                },
                "planoId": {
                    "type": "integer"
                },
                "responsavel": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                },
                "vencimento": {
                    "type": "string"
                }
            }
        },
        "model.Empresa": {
            "type": "object",
            "properties": {
                "cnpj": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "epis": {
                    "type": "integer"
                },
                "funcionarios": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "mensalidade": {
                    "type": "number"
                },
                "nome": {
                    "type": "string"
                },
                "observacoes": {
                    "type": "string"
                },
                "plano": {
                    "type": "string"
                },
                "responsavel": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                },
                "vencimento": {
                    "type": "string"
                }
            }
        },
        "model.EmpresaInserir": {
            "type": "object",
            "required": [
                "cnpj",
                "email",
                "nome_fantasia",
                "plano",
                "responsavel",
                "status",
                "vencimento"
            ],
            "properties": {
                "cnpj": {
                    "type": "string",
                    "maxLength": 40
                },
                "email": {
                    "type": "string",
                    "maxLength": 40
                },
                "nome_fantasia": {
                    "type": "string"
                },
                "observacoes": {
                    "type": "string",
                    "maxLength": 150
                },
                "plano": {
                    "type": "string"
                },
                "responsavel": {
                    "type": "string",
                    "maxLength": 40
                },
                "status": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string",
                    "maxLength": 40
                },
                "vencimento": {
                    "type": "string"
                }
            }
        },
        "model.EmpresaRecente": {
            "type": "object",
            "properties": {
                "epis": {
                    "type": "integer"
                },
                "funcionarios": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "mensalidade": {
                    "type": "number"
                },
                "nome": {
                    "type": "string"
                },
                "plano": {
                    "type": "string"
                },
                "responsavel": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subdominio": {
                    "type": "string"
                }
            }
        },
...
        "model.ResumoDashboard": {
            "type": "object",
            "properties": {
                "empresasAtivas": {
                    "type": "integer"
                },
                "empresasBloqueadas": {
                    "type": "integer"
                },
                "empresasEmTeste": {
                    "type": "integer"
                },
                "receitaMensal": {
                    "type": "number"
                },
                "totalEmpresas": {
                    "type": "integer"
                },
                "totalEntregas": {
                    "type": "integer"
                },
                "totalEpis": {
                    "type": "integer"
                },
                "totalFuncionarios": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}